refactor(matrix): extract matrix tools into tools/matrix.py module
Extract 470 lines of matrix operations (3 helpers, 5 MCP tools) from
server.py into src/math_mcp/tools/matrix.py using FastMCP mount()
composition pattern.

Changes:
- Create src/math_mcp/tools/ package with matrix.py module
- Move _check_numpy_available, _validate_matrix, _format_matrix helpers
- Move matrix_multiply, matrix_transpose, matrix_determinant,
  matrix_inverse, matrix_eigenvalues tools
- Use create_matrix_tools() factory pattern for tool registration
- Mount matrix_mcp sub-server with no prefix (preserves tool names)
- Move NUMPY_AVAILABLE flag to matrix module

Does not absorb #143; the raise-and-catch pattern for numpy availability
is simpler and consistent across all validation errors.

Part 2 of 4 for #140.
Closes #155.
